requirements.txt
Package                Version
---------------------- -----------
aiohappyeyeballs       2.6.1
aiohttp                3.11.16
aiohttp-retry          2.9.1
aiosignal              1.3.2
attrs                  25.3.0
blinker                1.9.0
certifi                2025.1.31
charset-normalizer     3.4.1
click                  8.1.8
Flask                  3.1.0
Flask-Login            0.6.3
frozenlist             1.6.0
h11                    0.14.0
idna                   3.10
iniconfig              2.1.0
itsdangerous           2.2.0
Jinja2                 3.1.6
MarkupSafe             3.0.2
multidict              6.4.3
mysql-connector-python 9.3.0
outcome                1.3.0.post0
packaging              24.2
pip                    25.0.1
pluggy                 1.5.0
propcache              0.3.1
PyJWT                  2.10.1
PySocks                1.7.1
pytest                 8.3.5
requests               2.32.3
selenium               4.31.0
sniffio                1.3.1
sortedcontainers       2.4.0
trio                   0.29.0
trio-websocket         0.12.2
twilio                 9.5.2
typing_extensions      4.13.2
urllib3                2.4.0
websocket-client       1.8.0
Werkzeug               3.1.3
wsproto                1.2.0
yarl                   1.20.0